re-organize config directory
