+++ paper/src/main/java/com/badbones69/crazycrates/api/objects/Prize.java
@@ -7,8 +7,8 @@ import com.badbones69.crazycrates.api.enums.misc.Keys;
 import com.badbones69.crazycrates.utils.ItemUtils;
 import com.badbones69.crazycrates.utils.MiscUtils;
 import com.badbones69.crazycrates.api.builders.ItemBuilder;
-import com.badbones69.crazycrates.common.config.ConfigManager;
-import com.badbones69.crazycrates.common.config.impl.messages.CrateKeys;
+import com.ryderbelserion.crazycrates.common.plugin.configs.ConfigManager;
+import com.ryderbelserion.crazycrates.common.plugin.configs.types.locale.CrateKeys;
 import com.ryderbelserion.vital.common.utils.math.MathUtil;
 import com.ryderbelserion.vital.paper.api.enums.Support;
 import com.ryderbelserion.vital.paper.util.AdvUtil;
