+++ paper/src/main/java/com/badbones69/crazycrates/commands/crates/types/admin/crates/migrator/ICrateMigrator.java
@@ -5,7 +5,7 @@ import com.badbones69.crazycrates.CrazyCrates;
 import com.badbones69.crazycrates.api.enums.Messages;
 import com.badbones69.crazycrates.utils.ItemUtils;
 import com.badbones69.crazycrates.commands.crates.types.admin.crates.migrator.enums.MigrationType;
-import com.badbones69.crazycrates.common.config.ConfigManager;
+import com.ryderbelserion.crazycrates.common.plugin.configs.ConfigManager;
 import com.badbones69.crazycrates.tasks.crates.CrateManager;
 import com.ryderbelserion.vital.paper.api.files.CustomFile;
 import com.ryderbelserion.vital.paper.api.files.FileManager;
