+++ paper/src/main/java/com/badbones69/crazycrates/commands/crates/types/admin/crates/migrator/types/plugins/ExcellentCratesMigrator.java
@@ -3,7 +3,7 @@ package com.badbones69.crazycrates.commands.crates.types.admin.crates.migrator.t
 import com.badbones69.crazycrates.api.enums.misc.Files;
 import com.badbones69.crazycrates.commands.crates.types.admin.crates.migrator.ICrateMigrator;
 import com.badbones69.crazycrates.commands.crates.types.admin.crates.migrator.enums.MigrationType;
-import com.badbones69.crazycrates.common.config.impl.ConfigKeys;
+import com.ryderbelserion.crazycrates.common.plugin.configs.types.config.ConfigKeys;
 import com.ryderbelserion.vital.paper.api.files.CustomFile;
 import com.ryderbelserion.vital.paper.util.AdvUtil;
 import com.ryderbelserion.vital.paper.util.ItemUtil;
